html/browsers/windows/embedded-opener-remove-frame.html

<!doctype html>
<title>opener and "removed" embedded documents</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<iframe name=matchesastring></iframe>
<script>
async_test(t => {
  const frame = document.querySelector("iframe"),
        frameW = frame.contentWindow;
  frame.onload = t.step_func(() => {
    // Firefox and Chrome/Safari load differently
    if (frame.contentWindow.location.href === "about:blank") {
      return;
    }

    // Test bits
    assert_equals(frameW.opener, window, "opener before removal");

    const openerDesc = Object.getOwnPropertyDescriptor(frameW, "opener"),
          openerGet = openerDesc.get;

    assert_equals(openerGet(), window, "opener before removal via directly invoking the getter");
    frame.remove();
    assert_equals(frameW.opener, null, "opener after removal");
    assert_equals(openerGet(), null, "opener after removal via directly invoking the getter");

    frameW.opener = "immaterial";
    assert_equals(openerGet(), null, "opener after setting it \"immaterial\" via directly invoking the getter");
    assert_equals(frameW.opener, "immaterial");

    t.done();
  });
  window.open("/common/blank.html", "matchesastring");
});
</script>
